Refactor dependency analysis and MCP server config
Extracted dependency analysis logic into a new dependency-analyzer module and updated get-dependents handler to use it, improving modularity and maintainability. Introduced server-config module to centralize MCP server configuration and handler registration, and refactored server.ts to use these abstractions. This change reduces code duplication and clarifies responsibilities across the MCP CLI codebase.
